# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: HEI Admin API
  description: |
    _Programmatically connect to a computer programming [school](https://hei.school)._
    After [joining us](mailto:contact@hei.school),
    you can get an identification token from
    our [dev](https://dev-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=5s8cg50doahmu855rlc8fr6qmp&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-dev.hei.school%2Fwhoami)
    or [prod](https://prod-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=i8bg538jpfu6mqmqb61m26trd&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-prod.hei.school%2Fwhoami)
    authentication service.
    Then, start playing with our system!

    The implementation of our API is [publicly disclosed](https://github.com/hei-school/hei-admin-api).
    You are welcome to try and compromise it. Happy hacking!
servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'
security:
  - BearerAuth: [ ]
paths:
...
  /staff_members:
    get:
      tags:
        - Users
      summary: Get all staff members
      operationId: getStaffMembers
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: status
          required: false
          in: query
          description: Filter staff members by status value return all by default
          schema:
            $ref: '#/components/schemas/EnableStatus'
        - name: sex
          required: false
          in: query
          description: Filter staff members by sex value return all by default
          schema:
            $ref: '#/components/schemas/Sex'
        - name: first_name
          in: query
          description: Filter staff members by first name, case is ignored
          schema:
            type: string
        - name: last_name
          in: query
          description: Filter staff members by last name, case is ignored
          schema:
            type: string
      responses:
        '200':
          description: List of staff members, ordered by ref.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMember'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Create or update staff members
      operationId: crupdateStaffMembers
      requestBody:
        description: staff members
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StaffMember'
      responses:
        '200':
          description: List of staff members, ordered by ref.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMember'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /staff_members/{id}:
    get:
      tags:
        - Users
      summary: Get staff member by the given id
      operationId: getStaffMemberById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The staff member
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/StaffMember'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Update staff member info
      operationId: updateStaffMember
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: staff member
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffMember'
      responses:
        '200':
          description: List of staff members, ordered by ref.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    UserIdentifier:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
          example: STD000001
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        nic:
          type: string
...
    CrupdateUser:
      allOf:
        - $ref: '#/components/schemas/UserIdentifier'
        - type: object
          properties:
            sex:
              $ref: '#/components/schemas/Sex'
            birth_date:
              type: string
              format: date
            birth_place:
              type: string
            address:
              type: string
            phone:
              type: string
            entrance_datetime:
              type: string
              format: date-time
            coordinates:
              $ref: '#/components/schemas/Coordinates'
            high_school_origin:
              type: string
            status:
              $ref: '#/components/schemas/EnableStatus'
...
    Sex:
      type: string
      enum:
        - M
        - F
    EnableStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
        - SUSPENDED
...
    Teacher:
      allOf:
        - $ref: '#/components/schemas/CrupdateUser'
      properties:
        profile_picture:
          type: string
...
    Coordinates:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
      default: 15
...
    StaffMember:
      allOf:
        - $ref: '#/components/schemas/Teacher'
      type: object
      properties:
        ending_service:
          type: string
          format: date-time
        cnaps:
          type: string
        ostie:
          type: string
        function:
          type: string
        degree:
          type: string
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error